+++ src/main/java/io/github/viciscat/kustweaks/mixin/tg/GenericProjectileAccessor.java
@@ -1,0 +1,13 @@
+package io.github.viciscat.kustweaks.mixin.tg;
+
+import net.minecraft.entity.EntityLivingBase;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+import techguns.entities.projectiles.GenericProjectile;
+
+@Mixin(GenericProjectile.class)
+public interface GenericProjectileAccessor {
+
+    @Accessor(value = "shooter", remap = false)
+    EntityLivingBase getShooter();
+}